Adjust the file architecture (#32)
* Adjust file architecture

* Split the util directory
