+++ backend/api/swagger/route.py
@@ -3,10 +3,10 @@ from typing import Any
 
 from flask import Blueprint, make_response
 
-swagger_bp = Blueprint("swagger", __name__)
+swagger_api_bp = Blueprint("swagger", __name__)
 
 
-@swagger_bp.route("/swagger_spec")
+@swagger_api_bp.route("/swagger_spec")
 def fetch_swagger_spec_route():
     swagger_spec: str = None
     with open("./swagger/spec.json") as f:
@@ -15,7 +15,7 @@ def fetch_swagger_spec_route():
     return make_response(swagger_spec_dict)
 
 
-@swagger_bp.route("/api")
+@swagger_api_bp.route("/api")
 def api_page_route():
     with open("./swagger/page.html") as f:
         return f.read()
